requirements.txt
openai
faiss-cpu
numpy
tqdm
python-dotenv
pdfplumber